requirements.txt
annotated-types==0.7.0
anyio==4.5.2
black==24.8.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
distro==1.9.0
exceptiongroup==1.2.2
fastapi==0.115.5
groq==0.12.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
Jinja2==3.1.4
lxml==5.3.0
MarkupSafe==2.1.5
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==10.4.0
platformdirs==4.3.6
pycparser==2.22
pydantic==2.10.1
pydantic_core==2.27.1
PyMuPDF==1.24.11
pypdfium2==4.30.0
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.17
sniffio==1.3.1
starlette==0.41.3
tomli==2.1.0
typing_extensions==4.12.2
uvicorn==0.32.1